// weather-forecast/src/types/intervalType.ts
interface intervalType {
  clouds: {
    all: number;
  };
  dt: number;
  dt_txt: string;
  main: {
    feels_like: number;
    grnd_level: number;
    humidity: number;
    pressure: number;
    sea_level: number;
    temp: number;
    temp_kf: number;
    temp_max: number;
    temp_min: number;
  };
  pop: number;
  sys: any;
  visibility: number;
  weather: any;
  wind: any;
  day?: any;
  rain?: any;
}

export default intervalType;