// weather-forecast/src/types/sortedInterval.ts
interface sortedIntervalType {
  date: any;
  dateName: string;
  dateNum: number;
  feelsTemp: number;
  hours: string;
  humidity: number;
  maxTemp: number;
  minTemp: number;
  rain: number;
  temp: number;
  time: number;
  weather: string;
  wind: number;
}

export default sortedIntervalType;